refactor(plugin): extract shared/borrow.rs (9× make_*_borrow helpers) (#1079)
Lifts the host-side β-shape borrow helpers
(make_pixel_buffer_borrow, make_storage_buffer_borrow,
make_uniform_buffer_borrow, make_texture_borrow,
make_vertex_buffer_borrow, make_index_buffer_borrow,
make_acceleration_structure_borrow, make_compute_kernel_borrow,
make_graphics_kernel_borrow) out of host_services/mod.rs into a
shared submodule. The vtable-accessor lookups (host_*_vtable())
still live in mod.rs and are pulled in via super:: — they'll
migrate to per-vtable submodules in subsequent commits.

Linux-only by construction (every helper was already
#[cfg(target_os = "linux")]). The submodule is module-gated rather
than per-function-gated.

Behavior identical; cargo test -p streamlib-engine --lib unchanged at
1407 passed / 2 pre-existing Deno failures / 126 ignored.
